refactor(assets and tailwind): change tailwindcss, and add file aliases
this pr changes tailwindcss from using output.css to using postcss and tailwind, it also adds
aliasing to reference certain files and cleans up the components a little bit
