+++ .eslintignore
@@ -3,3 +3,4 @@ babel.config.js
 .eslintrc.js
 lint-staged.config.js
 tailwind.config.js
+postcss.config.js
