+++ package.json
@@ -52,6 +52,7 @@
     "@typescript-eslint/eslint-plugin": "^5.25.0",
     "@typescript-eslint/parser": "^5.25.0",
     "@vitejs/plugin-react": "^1.3.0",
+    "autoprefixer": "^10.4.7",
     "babel-jest": "^28.1.0",
     "commitizen": "^4.2.4",
     "cz-conventional-changelog": "^3.3.0",
@@ -67,11 +68,13 @@
     "jsdom": "^19.0.0",
     "lint-staged": "^12.4.1",
     "npm-run-all": "^4.1.5",
+    "postcss": "^8.4.14",
     "prettier": "^2.6.2",
     "sort-package-json": "^1.57.0",
     "ts-jest": "^28.0.2",
     "typescript": "^4.6.3",
     "vite": "^2.9.9",
-    "vite-plugin-eslint": "^1.6.1"
+    "vite-plugin-eslint": "^1.6.1",
+    "vite-tsconfig-paths": "^3.4.1"
   }
 }
