+++ src/components/TopNav/index.tsx
@@ -1,0 +1,3 @@
+import TopNav from "./TopNav";
+
+export default TopNav;
