+++ src/main.tsx
@@ -1,6 +1,5 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-// Move this to using Index pattern
 import App from "./pages/App/App";
 import "./index.css";
