+++ tsconfig.json
@@ -1,23 +1,29 @@
 {
-  "compilerOptions": {
-    "types": ["jest", "vite/client"],
-    "target": "ESNext",
-    "useDefineForClassFields": true,
-    "lib": ["DOM", "DOM.Iterable", "ESNext"],
-    "allowJs": true,
-    "skipLibCheck": true,
-    "esModuleInterop": true,
-    "allowSyntheticDefaultImports": true,
-    "strict": true,
-    "forceConsistentCasingInFileNames": true,
-    "module": "ESNext",
-    "moduleResolution": "Node",
-    "resolveJsonModule": true,
-    "isolatedModules": true,
-    "noEmit": true,
-    "jsx": "react-jsx"
-  },
-  "include": ["**/*.ts", "**/*.tsx"],
-  "exclude": ["/node_modules/"],
-  "references": [{ "path": "./tsconfig.node.json" }]
+	"compilerOptions": {
+		"types": ["jest", "vite/client"],
+		"target": "ESNext",
+		"useDefineForClassFields": true,
+		"lib": ["DOM", "DOM.Iterable", "ESNext"],
+		"allowJs": true,
+		"skipLibCheck": true,
+		"esModuleInterop": true,
+		"allowSyntheticDefaultImports": true,
+		"strict": true,
+		"forceConsistentCasingInFileNames": true,
+		"module": "ESNext",
+		"moduleResolution": "Node",
+		"resolveJsonModule": true,
+		"isolatedModules": true,
+		"noEmit": true,
+		"jsx": "react-jsx",
+		"baseUrl": "./src",
+		"paths": {
+			"assets/*": ["/../assets/*"],
+			"components": ["/components/*"],
+			"pages": ["/pages/*"]
+		}
+	},
+	"include": ["**/*.ts", "**/*.tsx", "./src/*"],
+	"exclude": ["/node_modules/"],
+	"references": [{ "path": "./tsconfig.node.json" }]
 }
